services/auth/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.0.1
python-dateutil==2.8.2
Flask-Limiter==3.5.0
pytest==7.4.3
pytest-cov==4.1.0

